fix: 修复 Threshold int/double 类型不匹配 + 添加 XAML 编译诊断
- EdgeGestureSetting.Threshold 从 int 改为 double，解决 Slider.Value
  TwoWay x:Bind 类型不匹配导致 XAML 编译器静默失败
- 移除 Run.Text 的 x:Bind（double→string 在 Run 上可能有兼容问题）
- 添加 workflow 诊断步骤：构建失败时输出 XamlCompiler 的 output.json

+++ src/QuickLauncher/Models/SettingsModel.cs
@@ -43,7 +43,7 @@ public sealed class EdgeGestureSetting
 
     public bool Enabled { get; set; }
 
-    public int Threshold { get; set; } = 10;
+    public double Threshold { get; set; } = 10;
 
     public GestureAction Action { get; set; }
